refactor(lints): Pull out lint types
